requirements.txt
ollama>=0.3.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
xlrd>=2.0.0
xlwt>=1.3.0
xlutils>=2.0.0
openpyxl>=3.1.0
pypdf>=4.0.0
python-docx>=1.0.0
